spark/requirements.txt
pyspark
redis
boto3
pandas
python-dotenv
prometheus_client